+++ stringmatch/__init__.py
@@ -1,8 +1,7 @@
 # flake8: noqa
-from .exceptions import *
 from .match import *
 from .ratio import *
 from .strings import *
 
 __title__ = "stringmatch"
-__version__ = "0.4.1"
+__version__ = "0.5.0"