+++ stringmatch/match.py
@@ -1,12 +1,26 @@
 from typing import Optional
 
-from stringmatch.exceptions import EmptySearchException, InvalidLimitException
 from stringmatch.ratio import Ratio
 from stringmatch.strings import Strings
 
 
 class Match:
     """Contains methods for comparing and matching strings."""
+
+    def __init__(self, scorer: str = "levenshtein") -> None:
+        """Initialise the Match class with the correct scoring algorithm,
+        to be passed along to the Ratio class.
+
+        Parameters
+        ----------
+        scorer : str, optional
+            The scorer to use, by default "levenshtein"
+            Available scorers:
+                "levenshtein",
+                "jaro",
+                "jaro_winkler".
+        """
+        self.scorer = scorer
 
     def match(
         self,
@@ -18,11 +32,8 @@ class Match:
         ignore_case: bool = False,
         remove_punctuation: bool = False,
         only_letters: bool = False,
-        scorer: str = "levenshtein",
     ) -> bool:
         """Matches two strings, returns True if they are similar enough.
-        Be careful when using remove_punctuation and only_letters,
-        if they remove everything in the string, this will raise an EmptySearchException.
 
         Parameters
         ----------
@@ -33,29 +44,18 @@ class Match:
         score : int, optional
             The cutoff for the score, by default 70.
         latinise : bool, optional
-            If special unicode characters should be removed from the strings, by default False
+            If special unicode characters should be removed from the strings, by default False.
         ignore_case : bool, optional
-            If the strings should be compared ignoring case, by default False
+            If the strings should be compared ignoring case, by default False.
         remove_punctuation : bool, optional
-            If punctuation should be removed from the strings, by default False
+            If punctuation should be removed from the strings, by default False.
         only_letters : bool, optional
-            If the strings should only be compared by their latin letters, by default False
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
+            If the strings should only be compared by their latin letters, by default False.
 
         Returns
         -------
         bool
             If the strings are similar enough.
-
-        Raises
-        ------
-        EmptySearchException
-            If one of the strings to compare is empty.
         """
 
         if latinise:
@@ -76,12 +76,9 @@ class Match:
                 string2
             )
 
-        if not string1 or not string2:
-            raise EmptySearchException("Cannot compare an empty string.")
+        return Ratio(scorer=self.scorer).ratio(string1, string2) >= score
 
-        return Ratio().ratio(string1, string2, scorer=scorer) >= score
-
-    def match_with_score(
+    def match_with_ratio(
         self,
         string1: str,
         string2: str,
@@ -91,7 +88,6 @@ class Match:
         ignore_case: bool = False,
         remove_punctuation: bool = False,
         only_letters: bool = False,
-        scorer: str = "levenshtein",
     ) -> tuple[bool, int]:
         """Same as match, but returns the boolean in a tuple, together with the score.
 
@@ -104,29 +100,18 @@ class Match:
         score : int, optional
             The cutoff for the score, by default 70.
         latinise : bool, optional
-            If special unicode characters should be removed from the strings, by default False
+            If special unicode characters should be removed from the strings, by default False.
         ignore_case : bool, optional
-            If the strings should be compared ignoring case, by default False
+            If the strings should be compared ignoring case, by default False.
         remove_punctuation : bool, optional
-            If punctuation should be removed from the strings, by default False
+            If punctuation should be removed from the strings, by default False.
         only_letters : bool, optional
-            If the strings should only be compared by their latin letters, by default False
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
+            If the strings should only be compared by their latin letters, by default False.
 
         Returns
         -------
         tuple[bool, int]
             If the strings are similar and their score.
-
-        Raises
-        ------
-        EmptySearchException
-            If one of the strings to compare is empty.
         """
         kwargs = {
             "score": score,
@@ -134,12 +119,11 @@ class Match:
             "ignore_case": ignore_case,
             "remove_punctuation": remove_punctuation,
             "only_letters": only_letters,
-            "scorer": scorer,
         }
 
         return (
             self.match(string1, string2, **kwargs),
-            Ratio().ratio(string1, string2, scorer=scorer),
+            Ratio(scorer=self.scorer).ratio(string1, string2),
         )
 
     def get_best_match(
@@ -152,11 +136,8 @@ class Match:
         ignore_case: bool = False,
         remove_punctuation: bool = False,
         only_letters: bool = False,
-        scorer: str = "levenshtein",
     ) -> Optional[str]:
         """Returns the best match from a list of strings.
-        Be careful when using remove_punctuation and only_letters,
-        if they remove everything in the string, this will raise an EmptySearchException.
 
         Parameters
         ----------
@@ -165,21 +146,15 @@ class Match:
         string_list : list[str]
             The list of strings to compare to.
         score : int, optional
-            The cutoff for the score, by default 70
+            The cutoff for the score, by default 70.
         latinise : bool, optional
-            If special unicode characters should be removed from the strings, by default False
+            If special unicode characters should be removed from the strings, by default False.
         ignore_case : bool, optional
-            If the strings should be compared ignoring case, by default False
+            If the strings should be compared ignoring case, by default False.
         remove_punctuation : bool, optional
-            If punctuation should be removed from the strings, by default False
+            If punctuation should be removed from the strings, by default False.
         only_letters : bool, optional
-            If the strings should only be compared by their latin letters, by default False
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
+            If the strings should only be compared by their latin letters, by default False.
 
         Returns
         -------
@@ -192,16 +167,15 @@ class Match:
             "remove_punctuation": remove_punctuation,
             "ignore_case": ignore_case,
             "only_letters": only_letters,
-            "scorer": scorer,
         }
 
         return (
-            max(string_list, key=lambda s: Ratio().ratio(string, s, scorer=scorer))
+            max(string_list, key=lambda s: Ratio(scorer=self.scorer).ratio(string, s))
             if any(s for s in string_list if self.match(string, s, **kwargs))
             else None
         )
 
-    def get_best_match_with_score(
+    def get_best_match_with_ratio(
         self,
         string: str,
         string_list: list[str],
@@ -211,7 +185,6 @@ class Match:
         ignore_case: bool = False,
         remove_punctuation: bool = False,
         only_letters: bool = False,
-        scorer: str = "levenshtein",
     ) -> Optional[tuple[str, int]]:
         """Same as get_best_match, but returns a tuple with the best match and its score.
 
@@ -222,21 +195,15 @@ class Match:
         string_list : list[str]
             The list of strings to compare to.
         score : int, optional
-            The cutoff for the score, by default 70
+            The cutoff for the score, by default 70.
         latinise : bool, optional
-            If special unicode characters should be removed from the strings, by default False
+            If special unicode characters should be removed from the strings, by default False.
         ignore_case : bool, optional
-            If the strings should be compared ignoring case, by default False
+            If the strings should be compared ignoring case, by default False.
         remove_punctuation : bool, optional
-            If punctuation should be removed from the strings, by default False
+            If punctuation should be removed from the strings, by default False.
         only_letters : bool, optional
-            If the strings should only be compared by their latin letters, by default False
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
+            If the strings should only be compared by their latin letters, by default False.
 
         Returns
         -------
@@ -250,12 +217,13 @@ class Match:
             "remove_punctuation": remove_punctuation,
             "ignore_case": ignore_case,
             "only_letters": only_letters,
-            "scorer": scorer,
         }
 
         match = self.get_best_match(string, string_list, **kwargs)
 
-        return (match, Ratio().ratio(string, match, scorer=scorer)) if match else None
+        return (
+            (match, Ratio(scorer=self.scorer).ratio(string, match)) if match else None
+        )
 
     def get_best_matches(
         self,
@@ -263,19 +231,16 @@ class Match:
         string_list: list[str],
         *,
         score: int = 70,
-        limit: int = 5,
+        limit: Optional[int] = 5,
         latinise: bool = False,
         ignore_case: bool = False,
         remove_punctuation: bool = False,
         only_letters: bool = False,
-        scorer: str = "levenshtein",
     ) -> list[str]:
         """Matches a string to a list of strings, returns the strings found that are similar.
         If there are more than `limit` matches,
         only the `limit` best matches are returned, sorted by score.
         If no matches are found, returns an empty list.
-        Be careful when using remove_punctuation and only_letters,
-        if they remove everything in the string, this will raise an EmptySearchException.
 
         Parameters
         ----------
@@ -284,36 +249,27 @@ class Match:
         string_list : list[str]
             The list of strings to compare to.
         score : int, optional
-            The cutoff for the score, by default 70
+            The cutoff for the score, by default 70.
         limit : int, optional
-            The number of matches to return, by default 5
+            The number of matches to return, by default 5.
+            If you want to return every match, set this to 0 (or less than 0) or None.
         latinise : bool, optional
-            If special unicode characters should be removed from the strings, by default False
+            If special unicode characters should be removed from the strings, by default False.
         ignore_case : bool, optional
-            If the strings should be compared ignoring case, by default False
+            If the strings should be compared ignoring case, by default False.
         remove_punctuation : bool, optional
-            If punctuation should be removed from the strings, by default False
+            If punctuation should be removed from the strings, by default False.
         only_letters : bool, optional
-            If the strings should only be compared by their latin letters, by default False
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
+            If the strings should only be compared by their latin letters, by default False.
 
         Returns
         -------
         list[str]
             All of the matches found.
-
-        Raises
-        ------
-        InvalidLimitException
-            If the limit used is less than 1.
         """
-        if limit < 1:
-            raise InvalidLimitException("Limit must be greater than 1.")
+        # we return every match found if the limit is 0 or less
+        if limit is not None and limit < 1:
+            limit = None
 
         kwargs = {
             "score": score,
@@ -321,28 +277,26 @@ class Match:
             "remove_punctuation": remove_punctuation,
             "ignore_case": ignore_case,
             "only_letters": only_letters,
-            "scorer": scorer,
         }
 
         return sorted(
             [s for s in string_list if self.match(string, s, **kwargs)],
-            key=lambda s: Ratio().ratio(string, s, scorer=scorer),
+            key=lambda s: Ratio(scorer=self.scorer).ratio(string, s),
             # by default this would sort the list from lowest to highest.
             reverse=True,
         )[:limit]
 
-    def get_best_matches_with_score(
+    def get_best_matches_with_ratio(
         self,
         string: str,
         string_list: list[str],
         *,
         score: int = 70,
-        limit: int = 5,
+        limit: Optional[int] = 5,
         latinise: bool = False,
         ignore_case: bool = False,
         remove_punctuation: bool = False,
         only_letters: bool = False,
-        scorer: str = "levenshtein",
     ) -> list[tuple[str, int]]:
         """Same as get_best_matches, but returns a list of tuples with the best matches and their score.
 
@@ -353,33 +307,23 @@ class Match:
         string_list : list[str]
             The list of strings to compare to.
         score : int, optional
-            The cutoff for the score, by default 70
+            The cutoff for the score, by default 70.
         limit : int, optional
-            The number of matches to return, by default 5
+            The number of matches to return, by default 5.
+            If you want to return every match, set this to 0 (or less than 0) or None.
         latinise : bool, optional
-            If special unicode characters should be removed from the strings, by default False
+            If special unicode characters should be removed from the strings, by default False.
         ignore_case : bool, optional
-            If the strings should be compared ignoring case, by default False
+            If the strings should be compared ignoring case, by default False.
         remove_punctuation : bool, optional
-            If punctuation should be removed from the strings, by default False
+            If punctuation should be removed from the strings, by default False.
         only_letters : bool, optional
-            If the strings should only be compared by their latin letters, by default False
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
+            If the strings should only be compared by their latin letters, by default False.
 
         Returns
         -------
         list[tuple[str, int]]
             All of the matches found.
-
-        Raises
-        ------
-        InvalidLimitException
-            If the limit used is less than 1.
         """
         kwargs = {
             "score": score,
@@ -388,11 +332,10 @@ class Match:
             "remove_punctuation": remove_punctuation,
             "ignore_case": ignore_case,
             "only_letters": only_letters,
-            "scorer": scorer,
         }
 
         matches = self.get_best_matches(string, string_list, **kwargs)
 
         return [
-            (match, Ratio().ratio(string, match, scorer=scorer)) for match in matches
+            (match, Ratio(scorer=self.scorer).ratio(string, match)) for match in matches
         ]