+++ tests/test_ratio.py
@@ -1,6 +1,3 @@
-import pytest
-
-from stringmatch.exceptions import InvalidScorerException
 from stringmatch.ratio import Ratio
 
 
@@ -8,11 +5,13 @@ def test_ratio():
     assert Ratio().ratio("test", "test") == 100
     assert Ratio().ratio("bla", "nope") == 0
     assert Ratio().ratio("searchlib", "srechlib") == 82
-    assert Ratio().ratio("searchlib", "srechlib", scorer="jaro_winkler") == 93
-    assert Ratio().ratio("test", "th test", scorer="levenshtein") == 73
-    assert Ratio().ratio("test", "th test", scorer="jaro_winkler") == 60
-    with pytest.raises(InvalidScorerException):
-        assert Ratio().ratio("searchlib", "srechlib", scorer="nope")
+    assert Ratio(scorer="jaro_winkler").ratio("searchlib", "srechlib") == 93
+    assert Ratio(scorer="levenshtein").ratio("test", "th test") == 73
+    assert Ratio(scorer="jaro_winkler").ratio("test", "th test") == 60
+    assert Ratio(scorer="nope").ratio("searchlib", "srechlib") == 82
+    assert Ratio().ratio("", "f") == 0
+    assert Ratio("levenshtein").ratio_list("test", ["th test", "hwatever"]) == [73, 33]
+    assert Ratio("jaro_winkler").ratio_list("test", ["th test", "hwatever"]) == [60, 58]
 
 
 def test_ratio_list():